+++ pkg/azure/rbac_permissions_test.go
@@ -1,4 +1,4 @@
-package validators
+package azure
 
 import (
 	"reflect"
