+++ pkg/utils/azure/azure_test.go
